Add configurable services to send failure emails (#19)
* Add configurable services to send failure emails when email is set

Add a configuration option to define systemd services which should get an
OnFailure dependency, which in turn will send an email notifying the admin
about the error. The configuration is parsed by a systemd generator which
will add the dependency to the corresponding services.
If no email address is set in the configuration, then skip the creation of
the OnFailure drop-ins. The default configuration is changed to not contain
an email address any more, however for backwards compatibility set a
default in systemd-status-mail now if the configuration is empty.